Move subpackages into internal directory
The idea is to prevent exposing the packages as an
unintentional API surface.
