+++ internal/events/doc.go
