+++ internal/events/validate.go
