+++ internal/files/log.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/apex/log"
 
-	"github.com/atc0005/brick/events"
 	"github.com/atc0005/brick/internal/caller"
+	"github.com/atc0005/brick/internal/events"
 
 	"github.com/atc0005/go-ezproxy"
 )
