+++ internal/files/process.go
@@ -30,8 +30,8 @@ import (
 	"github.com/atc0005/go-ezproxy"
 	"github.com/atc0005/go-ezproxy/activefile"
 
-	"github.com/atc0005/brick/events"
 	"github.com/atc0005/brick/internal/caller"
+	"github.com/atc0005/brick/internal/events"
 	"github.com/atc0005/brick/internal/fileutils"
 )
